refactor(redux): migrate objects
